Course3

+++ src/main/java/com/jobex/jobapp/job/impl/JobServiceImpl.java
@@ -40,13 +40,13 @@ public class JobServiceImpl implements JobService {
 
     @Override
     public boolean deleteJobById(Long id) {
-        try{
+        if (jobRepository.existsById(id))
+        {
             jobRepository.deleteById(id);
             return true;
         }
-        catch(Exception e) {
-            return true;
-        }
+        else
+            return false;
     }
 
     @Override
